feat: rename debug command to connect and update related documentation
