+++ README.md
@@ -76,7 +76,7 @@ You can then modify your manifest files in the `k8s` folder as needed. To apply 
 For advanced debugging, connect to the cluster via the [K9s](https://k9scli.io) tool with:
 
 ```sh
-> ksail debug # to connect to the cluster
+> ksail connect # to connect to the cluster
 ```
 
 When you're done, you can stop the cluster to resume later:
