+++ tests/KSail.Tests.Unit/Commands/Connect/KSailConnectCommandTests.KSailDebugHelp_SucceedsAndPrintsIntroductionAndHelp.verified.txt
@@ -1,8 +1,8 @@
 ﻿Description:
-  Debug a cluster (❤️ K9s)
+  Connect to a cluster with K9s
 
 Usage:
-  testhost debug [options]
+  testhost connect [options]
 
 Options:
   -k, --kubeconfig <kubeconfig>  Path to kubeconfig file. [default: {UserProfile}/.kube/config]
