+++ tests/KSail.Tests.Unit/Commands/Root/KSailRootCommandTests.KSailHelp_SucceedsAndPrintsHelp.verified.txt
@@ -18,7 +18,7 @@ Commands:
   status    Show the status of a cluster
   list      List active clusters
   validate  Validate project files
-  debug     Debug a cluster (❤️ K9s)
+  connect   Connect to a cluster with K9s
   gen       Generate a resource
   secrets   Manage secrets
