+++ tests/KSail.Tests.Unit/Commands/Root/KSailRootCommandTests.KSail_SucceedsAndPrintsIntroduction.verified.txt
@@ -29,7 +29,7 @@ Commands:
   status    Show the status of a cluster
   list      List active clusters
   validate  Validate project files
-  debug     Debug a cluster (❤️ K9s)
+  connect   Connect to a cluster with K9s
   gen       Generate a resource
   secrets   Manage secrets
